==> 01_ETL_reviews.ipynb <==
### Importamos las librerias necesarias para comenzar el proceso de ETL 

import pandas as pd
import ast
import json
import numpy as np
from ast import literal_eval
import matplotlib.pyplot as plt
import seaborn as sns
from textblob import TextBlob
import herramientas
import warnings
warnings.filterwarnings("ignore")

***dataset: australian_user_reviews***


#### Extraccion de datos y primera exploracion

Se extraen los datos desde el archivo 'australian_user_reviews.json', se convierte en Dataframe y se observa su contenido.

#Lista para almacenar los diccionarios JSON de cada línea
review = []

#Ruta del archivo JSON
file_path = 'australian_user_reviews.json'

#Abrir el archivo y procesar cada línea
with open(file_path, 'r', encoding='utf-8') as file:
    for line in file:
        try:
            # Usar ast.literal_eval para convertir la línea en un diccionario
            json_data = ast.literal_eval(line)
            review.append(json_data)
        except ValueError as e:
            print(f"Error en la línea: {line}")
            continue

#Crear un DataFrame a partir de la lista de diccionarios
df_reviews = pd.DataFrame(review)
df_reviews

Se revisan los tipos de datos por columna y la cantidad de nulos.

herramientas.verifica_tipo_y_nulos(df_reviews)

Verificamos la existencia o no de duplicados por columnas user_id

filas_duplicadas = herramientas.verifica_duplicados_por_columna(df_reviews, 'user_id')
filas_duplicadas

Se observan 623 filas duplicadas en la columna 'user_id', pero se revisan si los review dentro de los datos anidados de 'df_review' la información se encuentra duplicada o si solo se duplica el 'user_id' porque hay mas de un comentario realizado por ese usuario

# Se revisa un usuario de ejemplo
user_id = '29123'
user_reviews = filas_duplicadas[filas_duplicadas['user_id'] == user_id]['reviews']

for review_list in user_reviews:
    for review in review_list:
        print(review['review'])
    print('-' * 40)

Se observa que los review son exactamente los mismos para cada registro, por lo que se decide borrar los duplicados, dejando la primer ocurrencia de los registros.

df_reviews = df_reviews.drop_duplicates(subset='user_id', keep='first')
herramientas.verifica_duplicados_por_columna(df_reviews, 'user_id')

Analizamos el tipo de dato de la columna 'reviews'

# Se observa el tipo de dato que contiene 'review'
df_reviews['reviews'][0]

Este dataset inicialmente contiene 3 columnas y 25799 filas, sin presencia de nulos. Las columnas que contiene son:
- **'user_id'**: identificador unico de usuario
- **'user_url'**: URL perfil del usuario
- **'reviews'**: Review de usuario en formato Json. El cual contiene una lista de diccionarios. Para cada usuario se tiene uno o mas diccionario con el review. Cada diccionario contiene:
    - **funny**: si alguien puso emoticón de gracioso al review.
    - **posted**: fecha de posteo del review en formato Posted April 21, 2011.
    - **last_edited**: fecha de la última edición.
    - **item_id**: identificador único del item (id del juego).
    - **helpful**: estadística donde otros usuarios indican si fue útil la información.
    - **recommend**: booleano que indica si el usuario recomienda o no el juego.
    - **review**: sentencia string con los comentarios sobre el juego.


#### Analisis de la columna 'reviews'

La columna 'reviews' esta anidada, siendo una lista con uno o mas diccionarios como elementos. Se generara una columna por cada diccionario para posteriormente hacer un registro por cada diccionario.

Este código toma un DataFrame con datos anidados en una columna ('reviews'), desanida esa columna para expandir sus elementos en filas separadas y luego expande esos elementos en múltiples columnas, finalmente concatenando estos nuevos datos con el DataFrame original.

#Tomamos los datos anidados en reviews , desanidamos la misma para expandir sus elementos en filas y multiples columnas para finalmente concarenarlas en el df_reviews
df_reviews1 = df_reviews.explode(['reviews'])
df_reviews2 = df_reviews1['reviews'].apply(pd.Series)
df_reviews3 = pd.concat([df_reviews1, df_reviews2], axis=1)

#Renombramos df_reviews3 por df_reviews
df_reviews = df_reviews3
#Visualizamos el dataset para verificar el paso anterior
df_reviews

Verificamos la cantidad de nulos 

herramientas.verifica_tipo_y_nulos(df_reviews)

Se eliminan las columnas que no necesitamos, en este caso serian las columnas: reviews, funny, last_edited, user_url. Luego visualizamos para verificar los cambios y guardamos en df_reviews

df_reviews_clean = df_reviews.drop('reviews', axis=1)
df_reviews_clean = df_reviews_clean.drop('funny', axis=1)
df_reviews_clean = df_reviews_clean.drop('last_edited', axis=1)
df_reviews_clean = df_reviews_clean.drop('user_url', axis=1)
df_reviews_clean = df_reviews_clean.drop(df_reviews_clean.columns[-1], axis=1)

df_reviews=df_reviews_clean 
df_reviews

Verificamos la existencia de valores nulos 

herramientas.verifica_tipo_y_nulos(df_reviews)

Eliminamos los valores nulos de las columnas: posted,  item_id, helpful, recommend, review	

df_reviews_clean = df_reviews.dropna(subset=['posted'])
df_reviews_clean = df_reviews_clean.dropna(subset=['item_id'])
df_reviews_clean = df_reviews_clean.dropna(subset=['helpful'])
df_reviews_clean = df_reviews_clean.dropna(subset=['recommend'])
df_reviews_clean = df_reviews_clean.dropna(subset=['review'])

df_reviews=df_reviews_clean 
df_reviews

Especificamos cada columna , que pertenece a ***dataset: australian_user_reviews*** , colocando 'reviews' antes del nombre de su columna 

df_reviews_clean = df_reviews_clean.add_prefix('reviews_')

df_reviews= df_reviews_clean
df_reviews

Se verifica si se eliminaron correctamente los valores nulos anteriormente encontrados y reveemos el tipo de dato

herramientas.verifica_tipo_y_nulos(df_reviews)

#### Modificacion del tipo de dato de ser necesario

- ***'reviews_posted'***

Se necesita que la fecha donde se hizo el posteo de la review este en formato YYYY-MM-DD. Por lo tanto, es necesario procesar la fecha y extraer los elementos relevantes. Se utilizará expresiones regulares para buscar y capturar los valores de año, mes y día dentro de la cadena de texto.

df_reviews['reviews_date'] = df_reviews['reviews_posted'].apply(herramientas.convertir_fecha)
df_reviews['reviews_date']

Se puede observar que hay 9932 registros que contienen un formato inválido distinto a los demas registros. En este caso, no contiene el año del posteo, pero con la función se imputó como 'Formato inválido'. Estos registros no se podrán consultar desde la API, pero las demás columnas serán útiles para aportar información.

df_reviews[df_reviews['reviews_date'] == 'Formato inválido']

Se elimina la columna reviews_posted , devido a que no aporta la informacion requerida 

df_reviews = df_reviews.drop('reviews_posted', axis=1)
df_reviews.columns

Se verifica si se elimino correctamente la columna 'reviews_posted'

herramientas.verifica_tipo_y_nulos(df_reviews)

df_reviews

### Análisis de sentimientos
Se pide crear una nueva columna llamada 'sentiment_analysis' que reemplace a 'reviews_review' donde se realice un análisis de sentimiento de los comentarios con la siguiente escala:

0 si es malo,
1 si es neutral o esta sin review
2 si es positivo.
Dado que el objetivo de este proyecto es realizar una prueba de concepto, consiguiendo un producto mínimo viable, se realiza un análisis de sentimiento básico utilizando TextBlob que es una biblioteca de procesamiento de lenguaje natural (NLP) en Python. El objetivo de esta metodología es asignar un valor numérico a un texto, en este caso a los comentarios que los usuarios dejaron para un juego determinado, para representar si el sentimiento expresado en el texto es negativo, neutral o positivo.


Se elimina los valores nulos de la columna 'reviews_review'

df_reviews_clean = df_reviews.dropna(subset=['reviews_review'])

df_reviews=df_reviews_clean 
df_reviews

Verifico valores nulos y tipo de dato

herramientas.verifica_tipo_y_nulos(df_reviews)

La metodología que utilizare es la siguiente: se toma una revisión de texto como entrada, utiliza TextBlob para calcular la polaridad de sentimiento y luego clasifica la revisión como negativa, neutral o positiva en función de la polaridad calculada. En este caso, se consideraron las polaridades por defecto del modelo, el cuál utiliza umbrales -0.2 y 0.2, siendo polaridades negativas por debajo de -0.2, positivas por encima de 0.2 y neutrales entre medio de ambos.

df_reviews['sentiment_analysis'] = df_reviews['reviews_review'].apply(herramientas.analisis_sentimiento)
df_reviews.head()

Se revisan algunos ejemplos para cada una de las clases de sentimiento.

herramientas.ejemplos_review_por_sentimiento(df_reviews['reviews_review'], df_reviews['sentiment_analysis'])

Creo la columna rating , que utilizare mas adelante

#Se aplica la función creada para agregar la columna 'ranting'
df_reviews['rating'] = df_reviews.apply(herramientas.columna_rating, axis=1)
df_reviews

Verifico nulos y tipos de datos

herramientas.verifica_tipo_y_nulos(df_reviews)

Observo los cambios

df_reviews

#### Carga del datasets: australian_user_reviews

Se guarda el conjunto de datos transformado como user_review_limpio.

df_reviews_limpio = 'data/user_review_limpio.csv'
df_reviews.to_csv(df_reviews_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {df_reviews_limpio}')
