==> 02_ETL_games.ipynb <==
### Importamos las librerias necesarias para comenzar el proceso de ETL 

import pandas as pd
import ast
import json
import numpy as np
from ast import literal_eval
import matplotlib.pyplot as plt
import seaborn as sns
from textblob import TextBlob
import herramientas
import warnings
warnings.filterwarnings("ignore")

***dataset: output_steam_games***

#### Extraccion de datos y primera exploracion

Se extraen los datos desde el archivo 'output_steam_games.json', se convierte en Dataframe y se observa su contenido.

#Leer el archivo línea por línea y cargar cada línea como un objeto JSON
with open('output_steam_games.json', 'r', encoding='utf-8') as archivo:
    data = [json.loads(line) for line in archivo]

#Convertir la lista de objetos JSON en un DataFrame
df_games = pd.DataFrame(data)
df_games

Se revisan los tipos de datos por columna y la cantidad de nulos.

herramientas.verifica_tipo_y_nulos(df_games)

Se observan filas que contienen vacíos en todo el registro. Se eliminan las mismas

df_games = df_games.dropna(how='all').reset_index(drop=True)
df_games.shape

Este conjunto contiene 13 columnas y quedó con 32135 filas luego de borrar las filas de NaN, aún contienen nulos en algunas columnas. Las columnas que contiene son:

- **publisher**: la empresa publicadora del contenido.
- **genres**: el género del item, es decir, del juego. Esta formado por una lista de uno o mas géneros por registro.
- **app_name**: el nombre del item, es decir, del juego.
- **title**:  el título del item.
- **url**:  la url del juego.
- **release_date**: la fecha de lanzamiento del item en formato 2018-01-04.
- **tags**:  la etiqueta del contenido. Esta formado por una lista de uno o mas etiquetas por registro.
- **reviews_url**:  la url donde se encuentra el review de ese juego.
- **specs**:  especificaciones de cada item. Es una lista con uno o mas string con las especificaciones.
- **price**:  el precio del item.
- **early_access**: indica el acceso temprano con un True/False.
- **id**:  el identificador único del contenido.
- **developer**: el desarrollador del contenido.

Se analizan si hay duplicados teniendo en cuenta la columna del id del item.

herramientas.verifica_duplicados_por_columna(df_games,'id')

Se observa que solo hay dos items duplicados. Por un lado el 'id' 612880 que se observa es el mismo registro en los dos casos. Por lo que se puede eliminar cualquier de ellos. Por otra parte, se observa que la segunda dupla de duplicados el 'id' es NaN y la primera aparición presenta mas cantidad de NaN que la segunda aparición.

Primero se busca por el 'developer' si este juego está ya registrado.

df_games[df_games['developer']=='Rocksteady Studios,Feral Interactive (Mac)']

Se puede ver que el item que no tiene 'id' es el 'id' 200260 y que el registro se encuentra completo. Por lo que se pueden borrar ambas filas dupliacadas.

# se eliminan los index deseados
filas_a_eliminar = [14573, 74, 30961]
df_games = df_games.drop(filas_a_eliminar)
# se verifica si hay dupliados
herramientas.verifica_duplicados_por_columna(df_games, 'id')

#### Modificacion del tipo de dato de ser necesario

- ***'release_date'***

Se necesita extraer el año de lanzamiento del item, para ello se hace una nueva columna con el dato si existe o con un "Dato no disponible" si no esta la fecha. Luego se elimina la columa 'release_date'.

# Se observan las cantidades de registros por cada fecha
df_games['release_date'].value_counts()

# Crea columna nueva con el año
df_games['release_anio'] = df_games['release_date'].apply(herramientas.extrae_anio)
# elimina la columna 'release_date'
df_games = df_games.drop('release_date', axis=1)
df_games.head()

df_games['release_anio'].unique()

- ***'publisher'***
- ***'app_name'***
- ***'title'*** 
- ***'developer'***

En estas columnas se observan valores nulos, por lo que se decide completarlos con valores del tipo 'Dato no disponible' para que el tipo de dato sea uno solo en la columna.

# SColumnas a transformar
columnas_a_completar = ['publisher', 'app_name', 'title', 'developer']
# Se rellenan los nulos
df_relleno = df_games[columnas_a_completar].fillna('Sin dato disponible')
# Se borran las columnas originales y se concatenan las rellenas con todo el dataframe
df_games = pd.concat([df_games.drop(columnas_a_completar, axis=1), df_relleno], axis=1)
df_games.head()

- ***'geners'***

La columna 'genres' esta formada por una lista de los distintos géneros de los videojuegos. Se necesita hacer una columna con cada género

df_games = df_games.explode('genres')
df_games = df_games.dropna(subset=['genres'])
df_games.head()

- ***'price'***

Se necesita trabajar con esta columna, pero hay valores de texto en casos de promociones o que el item es gratis. Por lo que se decide reemplazar esos valores por 0. Por otra parte, esta columna tiene algunos valores nulos, pero como mas adelante será necesario hacer operaciones, se adopta imputar como 0 los valores nulos.

df_games['price'] = df_games['price'].apply(herramientas.convierte_a_flotante)
df_games['price'].dtype

Se eliminan las columnas que no necesitamos, en este caso serian las columnas: 'publisher', 'url', 'app_name', 'specs', 'early_access', 'tags'. Luego visualizamos para verificar los cambios y guardamos en df_reviews

df_games_clean = df_games.drop('url', axis=1)
df_games_clean = df_games_clean.drop('specs', axis=1)
df_games_clean = df_games_clean.drop('tags', axis=1)
df_games_clean = df_games_clean.drop('reviews_url', axis=1)
df_games_clean = df_games_clean.drop('app_name', axis=1)

df_games = df_games_clean
df_games

herramientas.verifica_tipo_y_nulos(df_games)

df_games

#### Carga del datasets: output_steam_games

Se guarda el conjunto de datos transformado como steam_games_limpio

df_games_limpio = 'data/steam_games_limpio.csv'
df_games.to_csv(df_games_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {df_games_limpio}')
