==> 03_ETL_items.ipynb <==
### Importamos las librerias necesarias para comenzar el proceso de ETL 

import pandas as pd
import ast
import json
import numpy as np
from ast import literal_eval
import matplotlib.pyplot as plt
import seaborn as sns
from textblob import TextBlob
import herramientas
import warnings
warnings.filterwarnings("ignore")

***dataset: users_items***

#### Extraccion de datos y primera exploracion

Se extraen los datos desde el archivo 'australian_users_items.json', se convierte en Dataframe y se observa su contenido.

#Lista para almacenar los diccionarios JSON de cada línea
data_list = []

#Ruta del archivo JSON
file_path = 'australian_users_items.json'

#Abrir el archivo y procesar cada línea
with open(file_path, 'r', encoding='utf-8') as file:
    for line in file:
        try:
            # Usar ast.literal_eval para convertir la línea en un diccionario
            json_data = ast.literal_eval(line)
            data_list.append(json_data)
        except ValueError as e:
            print(f"Error en la línea: {line}")
            continue

#Crear un DataFrame a partir de la lista de diccionarios
df_items = pd.DataFrame(data_list)
df_items


Se revisan los tipos de datos por columna y la cantidad de nulos.

herramientas.verifica_tipo_y_nulos(df_items)

La columna 'items' es una lista, así que se explora la misma para conocer su estructura.

# se observa el tipo de dato en 'early_access'
df_items['items'][0]

Este conjunto contiene 5 columnas y 88309 filas, no tiene nulos. Las columnas que contiene son:

- **user_id**:  identificador único del usuario.
- **items_count**:  número entero que indica la cantidad de juegos que ha consumido el usuario.
- **steam_id**: número único para la plataforma.
- **user_url**: url del perfil del usuario
- **items**: lista de uno o mas diccionarios de los items que consume cada usuario. Cada diccionario tiene las siguientes claves:
    - **item_id**: identificados del item, es decir, del juego.
    - **item_name**: nombre del contenido que consume, es decir, del juego.
    - **playtime_forever**: tiempo acumulado que un usuario jugó a un juego.
    - **playtime_2weeks**: tiempo acumulado que un usuario jugó a un juego en las últimas dos semanas.

#Seguimos con la desanidacion de la estructura del Dataframe
df_items1 = df_items.explode(['items'])
df_items2 = pd.json_normalize(df_items1['items']).set_index(df_items1['items'].index)
df_items3= pd.concat([df_items2, df_items1], axis=1)
df_items = df_items3
df_items

herramientas.verifica_tipo_y_nulos(df_items)

Elimino las columnas que no utilizare posteriormente

#Realizamos la eliminacion de las columnas que no necesitamos
df_items_clean = df_items.drop('user_url', axis=1)
df_items_clean = df_items_clean.drop('playtime_2weeks', axis=1)
df_items_clean = df_items_clean.drop('items', axis=1)
#Renombramos nuevamente al nombre original
df_items = df_items_clean
#Visualizamos para verificar los cambios realizados
df_items

Verifico valores nulos

herramientas.verifica_tipo_y_nulos(df_items)

Eliminamos los valores nulos de las columnas 'item_id', 'item_name' y 'playtime_forever'

#Realizamos la limpieza de los valores "Nulos" en las columnas que requerimos
df_items_clean = df_items.dropna(subset=['item_id'])
df_items_clean = df_items_clean.dropna(subset=['item_name'])
df_items_clean = df_items_clean.dropna(subset=['playtime_forever'])

df_items=df_items_clean
df_items

Verifico tipo y nulos 

herramientas.verifica_tipo_y_nulos(df_items)

Verifico si poseen duplicados las columnas 'user_id', 'item_id', 'item_name', 'steam_id'. En caso de poseer los mismos , quiero saber el total 

df_items_repetidos_id = df_items[['user_id', 'item_id', 'item_name', 'steam_id']].duplicated().sum()
print("Cantidad de valores repetidos en la columna 'user_id' en relación con 'item_id', 'item_name' y 'steam_id':", df_items_repetidos_id)

Identifico los valores repetidos de 'user_id' en relacion con las demas columnas anteriormente mencionadas

#Encuentra los "user_id" repetidos en relacion con "item_id", "item_name" y "steam_id" y las veces que se repiten
df_items_repetidos_id  = df_items[['user_id', 'item_id', 'item_name', 'steam_id']][df_items[['user_id', 'item_id','item_name','steam_id']].duplicated(keep=False)]
df_items_repetidos_id_count = df_items_repetidos_id .value_counts()
#Filtra los "user_id" repetidos mayor o igual a 2 veces
df_items_repetidos_id = df_items_repetidos_id_count[df_items_repetidos_id_count >= 2].index
#Imprime los "user_id" repetidos y las veces que se repiten
df_items_repetidos_id_count

Elimino los valores repetidos finalmente

#Eliminar filas con "user_id", "item_id", "item_name" y "steam_id" repetidos y mantener la primera aparición
df_items_sin_repetidos = df_items.drop_duplicates(subset=['user_id','item_id','item_name','steam_id'], keep='first')
#Imprimir el DataFrame resultante
df_items_sin_repetidos

Guardos los cambios anteriormente realizados en el dataframe inicial 'df_items' y verifico si se guardaron correctamente 

df_items=df_items_sin_repetidos
df_items

Reviso si efectivamente existen duplicados

#Revisamos nuevamente la cantidad de valores repetidos en la columna 'user_id' en relacion con las columnas "item_id", "item_name" y "steam_id"
df_items_repetidos_id = df_items[['user_id', 'item_id', 'item_name','steam_id']].duplicated().sum()
print("Cantidad de valores repetidos en la columna 'user_id' en relacion con 'item_id', 'item_name' y 'steam_id':", df_items_repetidos_id)

Verifico la existencia de valores nulos 

herramientas.verifica_tipo_y_nulos(df_items)

Modificamos el tipo de dato de la columna 'playtime_forever'

df_items['playtime_forever']=df_items['playtime_forever'].astype(int)

Verificamos los cambios

herramientas.verifica_tipo_y_nulos(df_items)

df_items

#### Carga del datasets: australian_user_items

Se guarda el conjunto de datos transformado como user_items_limpio.

df_items_limpio = 'data/user_items_limpio.csv'
df_items.to_csv(df_items_limpio, index=False, encoding='utf-8')
print(f'Se guardó el archivo {df_items_limpio}')
