==> 05_Feature_Engineering.ipynb <==
## Feature Engineering

#### Importamos las librerias necesarias para realizar las funciones

import pandas as pd
import pyarrow as pa
import pyarrow.parquet as pq

%load_ext autoreload
%autoreload 2
import herramientas

import warnings
warnings.filterwarnings("ignore")

Leemos y guardamos en dataframes , las base de datos limpias 

df_reviews=pd.read_csv(r'data\user_review_limpio.csv')
df_items=pd.read_csv(r'data\user_items_limpio.csv')
df_games=pd.read_csv(r'data\steam_games_limpio.csv')

Verificamos nulos de las mismas 

herramientas.verifica_tipo_y_nulos(df_items)

herramientas.verifica_tipo_y_nulos(df_games)

herramientas.verifica_tipo_y_nulos(df_reviews)

Las columnas con presencias de nulos las limpiaremos en caso de utilizarlas , ademas posteriormente crearemos tablas por funciones optimizadas para luego convertirlas en .parquet para poder deployar en Render

- **Funcion 'PlayTimeGenre(genero)'**

Renombre las columas id por item_id en el df_games y le sacamos el prefijo 'reviews' del df_reviews, para posteriormente relacionar dicho dataframe con df_reviews

df_games.rename(columns={'id':'item_id'}, inplace= True)
df_reviews.rename(columns={'reviews_item_id':'item_id','reviews_user_id':'user_id'}, inplace= True)

Realizo el merge de los dataframes , y ademas le indicamos que nos traiga los distintos de 0 de la columna playtime_forever

df_merge=df_games[['genres','release_anio','item_id']]
df_merge1=df_items[['playtime_forever', 'item_id']]

df_merge1=df_merge1.loc[df_merge1['playtime_forever'] != 0]

df_merge2=pd.merge(df_merge,df_merge1,on='item_id')
df_merge2


Realizo la agrupacion por genres y release_anio.
Luego suma todos los valores en la columna ‘playtime_forever’ para cada grupo y reinicio el indice

df_grouped = df_merge2.groupby(['genres', 'release_anio'])['playtime_forever'].sum().reset_index()
df_grouped

Por ultimo guardo mi tabla en un archivo .parquet posteriormente avanzar con el proceso de deploy en render

#Guarda el DataFrame en el archivo PARQUET
df_grouped.to_parquet('data/df_playtime.parquet')

print(f'Se guardó el archivo {df_grouped}')

Leemos nuestro archivo parquet con un nuevo dataframe

df_play=pd.read_parquet('data/df_playtime.parquet')

Verificamos las dimenciones del mismo 

df_play

Creo mi funcion PlayTimeGenre( genero : str ): La cual debe devolver año con mas horas jugadas para dicho género.
Ejemplo de retorno: {"Año de lanzamiento con más horas jugadas para Género X" : 2013}

def PlayTimeGenre(genero):
      
    # Filtrar por el género especificado
    df_genre = df_play[df_play['genres'] == genero]
    
    # Si no hay datos para el género especificado, retorna un mensaje
    if df_genre.empty:
        return f"No hay datos para el género '{genero}'"
    
    # Agrupar por año y calcular las horas jugadas sumando los valores
    grouped = df_genre.groupby('release_anio')['playtime_forever'].sum()
    
    # Encontrar el año con más horas jugadas
    max_playtime_year = grouped.idxmax()
        
    # Retornar el resultado como un diccionario
    return {"Año de lanzamiento con más horas jugadas para Género {}".format(genero): max_playtime_year}

PlayTimeGenre('Action')

Finalmente comprobamos el funcionamiento de la funcion y seguimos con las funciones siguientes 

- **Funcion 'UserForGenre(genero)'**

Realizamos la creacion de mis dataframes necesarios para extraer la tabla necesaria para la consulta y posteriormente realizo el merge de las mismas para crear un dataframe

df_merge_g=df_games[['genres','item_id','release_anio',]]
df_merge_g1=df_items[['playtime_forever','item_id','user_id']]
df_merge_g2=pd.merge(df_merge_g1,df_merge_g,on= 'item_id')
df_merge_g2

Creo un dataframe que me traiga los valores distintos de 0 

df_merge_g2_clean=df_merge_g2.loc[df_merge_g2['playtime_forever'] != 0]
df_merge_g2_clean

Realizo la agrupoacion por las columas 'genres', 'user_id', 'release_anio' y sumo los valores en la columna ‘playtime_forever’, ademas de reestablecer el indice del dataframe

df_merge_g2_group=df_merge_g2_clean.groupby(['genres', 'user_id', 'release_anio'])['playtime_forever'].sum().reset_index()
df_merge_g2_group

Renombro el dataframe para guardar los cambios

df_useforgenre=df_merge_g2_group

#Guarda el DataFrame en el archivo PARQUET
df_useforgenre.to_parquet('data/df_useforgenre.parquet')

print(f'Se guardó el archivo {df_useforgenre}')

df_useforgenre=pd.read_parquet('data/df_useforgenre.parquet')
df_useforgenre

herramientas.verifica_tipo_y_nulos(df_useforgenre)

def UserForGenre(genero):
        
    # Filtrar por el género especificado
    df_genre = df_useforgenre[df_useforgenre['genres'] == genero]
    
    # Si no hay datos para el género especificado, retorna un mensaje
    if df_genre.empty:
        return f"No hay datos para el género '{genero}'"
    
    # Agrupar por usuario y género y calcular las horas jugadas sumando los valores
    grouped = df_genre.groupby(['user_id'])['playtime_forever'].sum()
    
    # Encontrar el usuario con más horas jugadas
    max_playtime_user = grouped.idxmax()
        
    # Agrupar por año y calcular las horas jugadas sumando los valores
    grouped_by_year = df_genre.groupby('release_anio')['playtime_forever'].sum()
    
    # Crear lista de acumulación de horas jugadas por año
    acumulacion_horas = [{'Año': year, 'Horas': hours} for year, hours in grouped_by_year.items()]
    
    # Retornar el resultado como un diccionario
    return {"Usuario con más horas jugadas para Género {}".format(genero): max_playtime_user, "Horas jugadas": acumulacion_horas}


UserForGenre('Action')

- **Funcion UsersRecommend(año)**

herramientas.verifica_tipo_y_nulos(df_reviews)

herramientas.verifica_tipo_y_nulos(df_games)

Verifico la existencia de valores = 2 en la column a 'rating'

df_reviews['rating'] == 2

# Convierte la columna 'reviews_date' a datetime
df_reviews['reviews_date'] = pd.to_datetime(df_reviews['reviews_date'], errors='coerce')

# Elimina las filas donde 'reviews_date' es NaT
df_reviews = df_reviews.dropna(subset=['reviews_date'])

# Extrae el año de 'reviews_date' a una nueva columna 'year'
df_reviews['year'] = df_reviews['reviews_date'].dt.year.astype(int)

# Elimina la columna 'reviews_date'
df_reviews = df_reviews.drop('reviews_date', axis=1)

df_reviews


df_reviews1 = df_reviews[df_reviews['year'].duplicated()]
df_reviews1


herramientas.verifica_tipo_y_nulos(df_reviews)

df_merge_r2 = pd.merge(df_games[['title','item_id']], df_reviews[['rating','item_id','year']], on='item_id')
df_merge_reco1 = df_merge_r2.dropna(subset=['rating'])
df_merge_reco1 = df_merge_reco1[df_merge_reco1['rating'] == 2]


df_merge_reco1

#Agrupar por año y título, contar la frecuencia y clasificar en orden descendente
df_grouped = df_merge_reco1.groupby(['year', 'title']).size().reset_index(name='count').sort_values(by=['year', 'count'], ascending=[True, False])

#Crear una nueva columna 'rank' dentro de cada grupo
df_grouped['rank'] = df_grouped.groupby('year')['count'].rank(ascending=False, method='first')

#Filtrar los tres títulos principales por año
df_top3 = df_grouped[df_grouped['rank'] <= 3]
df_top3

#Puedes seleccionar las columnas que desees en el nuevo DataFrame
result_df = df_top3[['year', 'title', 'rank']]
result_df

df_recom=result_df

def UsersRecommend(año: int):
    # Filtra el DataFrame df_recom por el año proporcionado y lo guarda en top3_by_year
    top3_by_year = df_recom[df_recom['year'] == año]

    # Inicializa una lista vacía para almacenar los resultados
    resultado = []
    
    # Itera sobre las filas del DataFrame top3_by_year
    for index, row in top3_by_year.iterrows():
        # Obtiene el puesto, el título y el año de la fila actual
        puesto = row['rank']
        titulo = row['title']
        año = int(row['year'])
        
        # Agrega un diccionario a la lista de resultados con la estructura {f"Puesto {puesto}": f"{titulo}"}
        resultado.append({f"Puesto {puesto}": f"{titulo}"})
    
    # Devuelve la lista de resultados
    return resultado


UsersRecommend(2015)

#Guarda el DataFrame en el archivo PARQUET
df_recom.to_parquet('data/df_recom.parquet')

print(f'Se guardó el archivo {df_recom}')

herramientas.verifica_tipo_y_nulos(df_merge_reco1)

Funcion **UsersWorstDeveloper(año)**

# Crea un nuevo DataFrame df_games_w que contiene solo las columnas 'developer' e 'item_id' de df_games
df_games_w=df_games[['developer','item_id']]

# Crea otro DataFrame df_reviews_w que contiene solo las columnas 'rating', 'item_id' y 'year' de df_reviews
df_reviews_w=df_reviews[['rating','item_id','year']]

# Une df_games_w y df_reviews_w en un nuevo DataFrame df_merge basándose en la columna 'item_id'
df_merge=pd.merge(df_games_w,df_reviews_w, on='item_id')

# Elimina las filas de df_merge donde la columna 'rating' es NaN y guarda el DataFrame resultante en df_merge_clean
df_merge_clean=df_merge.dropna(subset='rating')

# Filtra df_merge_clean para incluir solo las filas donde la columna 'rating' es igual a 1 y guarda el DataFrame resultante en df_merge_rating
df_merge_rating=df_merge_clean[df_merge_clean['rating'] == 1]

# Agrupa df_merge_rating por 'year' y 'developer', cuenta el número de filas en cada grupo, ordena los grupos por 'year' y 'count' en orden ascendente y guarda el DataFrame resultante en df_merge_w
df_merge_w = df_merge_rating.groupby(['year', 'developer']).size().reset_index(name='count').sort_values(by=['year', 'count'], ascending=[True, True])

# Crea una nueva columna 'rank' en df_merge_w que contiene el rango de 'count' dentro de cada grupo de 'year'
df_merge_w ['rank'] = df_merge_w .groupby('year')['count'].rank(ascending=False, method='first')

# Filtra df_merge_w para incluir solo las filas donde la columna 'rank' es menor o igual a 3 y guarda el DataFrame resultante en df_worst
df_worst = df_merge_w[df_merge_w['rank'] <= 3]

# Filtra df_worst para excluir las filas donde la columna 'year' es igual a 'Dato no disponible' y guarda el DataFrame resultante en df_worst_1
df_worst_1= df_worst[df_worst['year'] != 'Dato no disponible']

# Convierte la columna 'rank' en df_worst_1 a enteros
df_worst_1['rank']=df_worst_1['rank'].astype(int)
df_worst_1


herramientas.verifica_tipo_y_nulos(df_worst_1)

#Guarda el DataFrame en el archivo PARQUET
df_worst_1.to_parquet('data/df_worst.parquet')

print(f'Se guardó el archivo {df_worst_1}')

df_worst=pd.read_parquet('data/df_worst.parquet')

def UsersWorstDeveloper( año : int ):
    # Filtrar el DataFrame df_developer por el año proporcionado
    developer_by_year = df_worst_1[df_worst_1['year'] == año]

    # Obtener el top 3 de desarrolladoras con juegos MENOS recomendados y sus valores según rank
    top3_worst_developer = developer_by_year.sort_values(by='rank', ascending=True).head(3)

    # Formatear el resultado como lista de diccionarios
    result = [{"Puesto {}: {}".format(rank, developer)} for rank, developer in zip(top3_worst_developer['rank'], top3_worst_developer['developer'])]

    return result


UsersWorstDeveloper(2011)

- **Funcion: sentiment_analysis(año)**

# Selecciona las columnas 'developer' e 'item_id' del dataframe df_games y las guarda en un nuevo dataframe df_games_d
df_games_d=df_games[['developer','item_id']]

# Selecciona las columnas 'item_id', 'reviews_recommend' y 'sentiment_analysis' del dataframe df_reviews y las guarda en un nuevo dataframe df_reviews_s
df_reviews_s=df_reviews[['item_id','reviews_recommend','sentiment_analysis']]

# Une los dataframes df_games_d y df_reviews_s en base a la columna 'item_id'
df_merge=pd.merge(df_games_d,df_reviews_s, on='item_id')

# Agrupa el dataframe unido por las columnas 'developer' y 'sentiment_analysis', cuenta el número de 'reviews_recommend' para cada grupo, y resetea el índice del dataframe
df_merge_group=df_merge.groupby(['developer', 'sentiment_analysis'])['reviews_recommend'].count().reset_index() 

# Renombra la columna 'reviews_recommend' a 'reviews_recommend_count' en el dataframe df_merge_group
df_merge_group = df_merge_group.rename(columns={'reviews_recommend': 'reviews_recommend_count'})

# Guarda el dataframe df_merge_group en una nueva variable df_senti
df_senti=df_merge_group

# Muestra el dataframe df_senti
df_senti


herramientas.verifica_tipo_y_nulos(df_senti)

#Guarda el DataFrame en el archivo PARQUET
df_senti.to_parquet('data/df_senti.parquet')

print(f'Se guardó el archivo {df_senti}')

df_senti=pd.read_parquet('data/df_senti.parquet')
df_senti

La función sentiment_analysis toma como entrada un nombre de empresa desarrolladora de juegos y devuelve un diccionario que contiene la cantidad de reseñas negativas, neutrales y positivas para esa empresa.

def sentiment_analysis(empresa_desarrolladora: str): 
    # Filtra el DataFrame df_senti por la empresa desarrolladora proporcionada y lo guarda en developer_df
    developer_df = df_senti[df_senti['developer'] == empresa_desarrolladora]

    # Crea un diccionario result con la estructura {empresa_desarrolladora: {'Negative': 0, 'Neutral': 0, 'Positive': 0}}
    result = {empresa_desarrolladora: {'Negative': 0, 'Neutral': 0, 'Positive': 0}}

    # Itera sobre las columnas 'sentiment_analysis' y 'reviews_recommend_count' del DataFrame developer_df
    for sentiment, count in zip(developer_df['sentiment_analysis'], developer_df['reviews_recommend_count']):
        # Crea un mapeo entre los valores numéricos de la columna 'sentiment_analysis' y las categorías de sentimiento correspondientes
        sentiment_mapping = {0: 'Negative', 1: 'Neutral', 2: 'Positive'}
        
        # Usa el mapeo para obtener la categoría de sentimiento correspondiente al valor numérico
        sentiment_category = sentiment_mapping[sentiment]
        
        # Suma la cantidad de reseñas (count) a la categoría de sentimiento correspondiente en el diccionario result
        result[empresa_desarrolladora][sentiment_category] += count

    # Devuelve el diccionario result
    return result


sentiment_analysis('Valve')
