==> 06_Modelo_Recomendacion.ipynb <==
## Modelo de Recomendacion

### Importamos las librerias necesarias para comenzar el proceso de ETL 

import pandas as pd
import numpy as np
import herramientas
import scipy as sp
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import CountVectorizer
import operator


import pyarrow as pa
import pyarrow.parquet as pq

Se leen los datos del archivo .csv y se guardan en un dataframe

df_games=pd.read_csv(r'data\steam_games_limpio.csv')
df_items=pd.read_csv(r'data\user_items_limpio.csv')
df_items

Verificamos la existencia de nulos y los tipos de datos de las columnas de los dataframes 

herramientas.verifica_tipo_y_nulos(df_games)

herramientas.verifica_tipo_y_nulos(df_items)

Verifico duplicados 

df_games_m=df_games[['developer','id', 'title']]
df_games_m

Agrupo por titulo y aplico una función lambda a la columna ‘developer’. Esta función convierte cada entrada en la columna ‘developer’ a una cadena (si no lo es ya) y luego une todas las cadenas en el grupo con ', ’ entre ellas. Esto resulta en una única cadena que contiene todos los valores de ‘developer’ para cada título, separados por comas

#Se agrupa por título y concatena las entradas de la columna 'rating'
df_agrupado_con_developer = df_games_m.groupby('title')['developer'].apply(lambda x: ', '.join(x.astype(str))).reset_index()

#Se visualiza la tabla
df_agrupado_con_developer 

Verifico el tipo de dato y si existen nulos 

herramientas.verifica_tipo_y_nulos(df_agrupado_con_developer)

Verifico la existencia de duplicados en todo el dataframe

df_agrupado_con_developer.duplicated().any()

Crea un objeto CountVectorizer, que es una herramienta para convertir una colección de documentos de texto en una matriz de conteos de tokens. 

- **max_features = 2000**: Limita el número de características (palabras únicas) que el vectorizador tomará en cuenta, manteniendo solo las 2000 palabras más frecuentes.
- **stop_words = "english"**: Le dice al vectorizador que ignore las palabras comunes en inglés (como ‘the’, ‘is’, ‘and’, etc.) que no aportan mucha información sobre el contenido del texto.

Luego, el método **fit_transform** se aplica a la columna “developer”. Este método aprende el vocabulario de la colección de documentos (en este caso, los nombres de los desarrolladores) y devuelve una matriz donde cada fila corresponde a un documento y cada columna a una palabra del vocabulario. Las entradas de la matriz son el número de veces que cada palabra aparece en cada documento.

El método toarray convierte la 'matriz dispersa' devuelta por **fit_transform** en una 'matriz densa' (un array de NumPy), y shape devuelve las dimensiones de esta matriz. La primera dimensión es el número de documentos (en este caso, el número de filas en df_agrupado_con_developer) y la segunda dimensión es el número de palabras en el vocabulario (hasta un máximo de 2000, como se especificó anteriormente).

cv = CountVectorizer(max_features = 2000, stop_words="english")
cv.fit_transform(df_agrupado_con_developer["developer"]).toarray().shape

Este método **(vectors = cv.fit_transform(df_agrupado_con_developer["developer"]))** aprende el vocabulario de la colección de documentos (en este caso, los nombres de los desarrolladores) y devuelve una matriz donde cada fila corresponde a un documento y cada columna a una palabra del vocabulario. Las entradas de la matriz son el número de veces que cada palabra aparece en cada documento.

 - **.toarray()**: Convierte la matriz dispersa devuelta por fit_transform en una matriz densa (un array de NumPy). 



vectors = cv.fit_transform(df_agrupado_con_developer["developer"]).toarray()
vectors

Esta función **(cosine_similarity(vectors))** calcula la similitud del coseno entre cada par de vectores en la matriz vectors. En este caso guarda la matriz de similitud del coseno en la variable similarity

- **similarity[0]**: Muestra el primer elemento de la matriz de similitud del coseno, que es un vector que contiene las similitudes del coseno entre el primer vector en vectors y todos los demás vectores.


similarity = cosine_similarity(vectors)
similarity[0]

Por medio de la funcion **sorted** se obtendra una lista de las cinco tuplas que representan los cinco vectores más similares al primer vector, junto con sus índices originales y sus valores de similitud del coseno.

sorted(list(enumerate(similarity[0])), reverse=True, key=lambda x:x[1])[1:6]

La siguiente funcion (recomendacion) toma como entrada el nombre de un juego y devuelve una lista de juegos recomendados basándose en la similitud de los desarrolladores.

def recomendacion(game):
    indice_juegos = df_agrupado_con_developer[df_agrupado_con_developer["title"]==game].index[0]
    distances = similarity[indice_juegos]
    lista_juegos = sorted(list(enumerate(distances)), reverse=True, key=lambda x: x[1])[1:6]
    recommended_titles = [df_agrupado_con_developer.iloc[i[0]]['title'] for i in lista_juegos]
    return recommended_titles

Aplico la función recomendacion a la columna ‘title’ del dataframe df_agrupado_con_developer y almacenando los resultados en una nueva columna llamada ‘model’.

Luego de ejecutar este código, cada fila en df_agrupado_con_developer tendrá una nueva columna ‘model’ que contiene una lista de juegos recomendados para el título de esa fila.


df_agrupado_con_developer['model']=df_agrupado_con_developer['title'].apply(recomendacion)

Elimino  la columna 'developer '

df_agrupado_con_developer_del=df_agrupado_con_developer.drop('developer', axis=1, inplace= True)

Creamos un dataframe el cual nos traiga las columas 'id', y 'title' del dataframe df_games. Luego verificamos los valores duplicados del mismo

df_games_m_m=df_games_m[['id','title']]
df_games_clean=df_games_m_m.drop_duplicates(subset='title',keep='first')
df_games_clean

Verifico la existencia de los duplicados de la columna 'title'

herramientas.verifica_duplicados_por_columna(df_games_clean,'title')

Verifico la existencia de valores nulos y el tipo de dato 

herramientas.verifica_tipo_y_nulos(df_games_clean)

Realizo un merge entre los dataframes, por la columna en comun 'title'

df_merge=pd.merge(df_agrupado_con_developer,df_games_clean, on='title', how='inner')
df_merge

Verifico la existencia de nulos 

herramientas.verifica_tipo_y_nulos(df_merge)

Del nuevo dataframe extraigo solo las columnas 'id' y 'model'

df_merge_id=df_merge[['id','model']]
df_merge_id

Guardamos el ultimo dataframe en un archivo parquet, el cual utilizaremos para realizar la consulta

#Guarda el DataFrame en el archivo PARQUET
df_merge_id.to_parquet('data/df_recomendacion.parquet')

print(f'Se guardó el archivo {df_merge_id}')

- **Funcion: recomendacion_juego (id)**

La función recomendacion_juego(id:int) Ingresando el id de producto, deberíamos recibir una lista con 5 juegos recomendados similares al ingresado.

def recomendacion_juego(id:int):
    # Filtra el DataFrame df_merge_id por el id proporcionado y obtiene el 'model' correspondiente
    modelo = df_merge_id[df_merge_id['id'] == id]['model'].iloc[0]
    
    # Inicializa un diccionario vacío para almacenar las recomendaciones
    recomendaciones_dict = {}
    
    # Si el modelo tiene elementos, procede a llenar el diccionario de recomendaciones
    if len(modelo) > 0:
        for i in range(len(modelo)):
            # Agrega cada elemento del modelo al diccionario de recomendaciones
            recomendaciones_dict[i + 1] = modelo[i]
        # Devuelve el diccionario de recomendaciones
        return recomendaciones_dict
    else:
        # Si el modelo no tiene elementos, devuelve un mensaje de error
        return f"No se encontró un modelo para el id {id}"


recomendacion_juego(454790)

Verifico la existencia de alguno de los juegos dentro del dataframe para verificar el correcto funcionamiento del modelo 

df=df_merge[df_merge['title']=="My Lady"]
df
